SOLR-13151: apply SOLR-12148 patch
